Refactored Users API, updated database to v3.1, added history actions and controllers, fixed merge conflicts

+++ Backend/PicturesAPI/app/models/face_shape_link.py
@@ -4,7 +4,7 @@ Author: Diego C. <20026893@tafe.wa.edu.au>
 Created at: 29/09/2020 4:07 pm
 File: face_shape_link.py
 """
-from sqlalchemy import func, Column, String, Integer, DateTime, ForeignKey
+from sqlalchemy import func, Column, String, BIGINT, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 from app.database.db import Base
 
@@ -13,8 +13,8 @@ from .face_shape import FaceShape
 
 class FaceShapeLinks(Base):
     __tablename__ = "face_shape_links"
-    id = Column(Integer, primary_key=True, index=True)
-    face_shape_id = Column(Integer, ForeignKey("face_shapes.id"), nullable=False)
+    id = Column(BIGINT, primary_key=True, index=True)
+    face_shape_id = Column(BIGINT, ForeignKey("face_shapes.id"), nullable=False)
     link_name = Column(String(128), nullable=False, server_default="** ERROR: missing category **")
     link_url = Column(String(512), nullable=False, server_default="** ERROR: missing category **")
     date_created = Column(DateTime(timezone=True), server_default=func.now())
